Change return value of I2S:: availableForWrite to match new Print API

+++ libraries/I2S/src/I2S.h
@@ -51,7 +51,7 @@ public:
   virtual size_t write(uint8_t);
   virtual size_t write(const uint8_t *buffer, size_t size);
 
-  virtual size_t availableForWrite();
+  virtual int availableForWrite();
 
   int read(void* buffer, size_t size);
